fix(parser): preserve single-parse recovery contract
Move editor repair planning into the syntax layer and build one recoverable candidate before invoking OXC, keeping parser recovery framework-neutral and single-pass.
